refactor: move editEvent epics and actions to ViewEvent page
