+++ src/core_modules/capture-core/components/Pages/ViewEvent/ViewEventComponent/editEvent.actions.js
