+++ src/core_modules/capture-core/reducers/descriptions/editEvent.reducerDescription.js
@@ -1,6 +1,6 @@
 // @flow
 import { createReducerDescription } from '../../trackerRedux/trackerReducer';
-import { actionTypes as editEventActionTypes } from '../../components/Pages/EditEvent/editEvent.actions';
+import { actionTypes as editEventActionTypes } from '../../components/Pages/ViewEvent/ViewEventComponent/editEvent.actions';
 import { actionTypes as editEventDataEntryActionTypes } from '../../components/Pages/EditEvent/DataEntry/editEventDataEntry.actions';
 
 export const editEventPageDesc = createReducerDescription({
