+++ src/core_modules/capture-core/reducers/descriptions/mainPage.reducerDescription.js
@@ -1,6 +1,6 @@
 // @flow
 import { createReducerDescription } from '../../trackerRedux/trackerReducer';
-import { actionTypes as editEventPageUrlActionTypes } from '../../components/Pages/EditEvent/editEvent.actions';
+import { actionTypes as editEventPageUrlActionTypes } from '../../components/Pages/ViewEvent/ViewEventComponent/editEvent.actions';
 
 export const mainPageDesc = createReducerDescription({
     [editEventPageUrlActionTypes.EDIT_EVENT_FROM_URL]: (state) => {
